pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xhs-creator"
version = "1.1.0"
description = "AI 驱动的小红书内容创作 CLI 工具"
requires-python = ">=3.8"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "Pillow>=9.0",
]

[project.scripts]
xhs-creator = "xhs_creator.cli:cli"
